requirements.txt
streamlit
langchain-community
langchain-text-splitters
langchain-chroma
zhipuai
pypdf
cryptography
python-dotenv